+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-08-16 21:35:45
-// Commit: feat(core): implement shimmer loading skeleton in core
+// Updated at: 2026-08-18 13:09:55
+// Commit: test(core): implement widget integration tests for core
 
 class coreComponentState {
-  final String commitRef = '3658';
-  final String lastMessage = 'feat(core): implement shimmer loading skeleton in core';
-  final DateTime timestamp = DateTime.parse('2026-08-16T21:35:45+05:00');
+  final String commitRef = '3692';
+  final String lastMessage = 'test(core): implement widget integration tests for core';
+  final DateTime timestamp = DateTime.parse('2026-08-18T13:09:55+05:00');
 }